Version Beta 1.1.1

+++ cogs/fun.py
@@ -40,15 +40,16 @@ class Fun(commands.Cog):
         gezogene_karte = "Du hast folgende Karte gezogen: **" + karten[random.randint(0, 31)] + "**"
         await ctx.send(gezogene_karte)
 
-    @commands.command()
+    @commands.Cog.listener()
     async def on_message(self, message):
-        if (message.content.startswith("print('") and message.count("'") == 2 and
-            message.endswith("')")) or (message.content.startswith('print("') and message.count('")') == 2
-                                        and message.endswith("'")):
+        if bp.user(message.author) and ((message.content.startswith("py.print('") and message.content.count("'") == 2
+                                         and message.content.endswith("')")) or
+                                        (message.content.startswith('py.print("') and message.content.count('"') == 2
+                                         and message.content.endswith('")'))):
             content = message.content[:-2]
-            content = content[:7]
+            content = content[10:]
             await bp.delete_cmd(message)
-            await message.send(f'{message.author} sagt {content}')
+            await message.channel.send(f'{message.author} sagt {content}')
 
 
 def setup(client):
